rename utility to minimal_output, sort metric names
