+++ test/adaptive_load/minimal_output.h
